Internalize round_after_divisor
